requirements.txt
asgiref==3.5.2
backports.zoneinfo==0.2.1
certifi==2022.9.24
charset-normalizer==2.1.1
Django==4.1.3
idna==3.4
PyJWT==2.6.0
python-dotenv==0.21.0
pytz==2022.6
requests==2.28.1
sqlparse==0.4.3
twilio==7.15.4
urllib3==1.26.13
